refactor: decompose dashboard server hotspots (stage 1)
