+++ dashboard/tests/unit/maintainability/file-size-guard.test.ts
@@ -1,0 +1,31 @@
+import fs from 'node:fs';
+import path from 'node:path';
+
+import { describe, it, expect } from 'vitest';
+
+const DASHBOARD_ROOT = path.resolve(import.meta.dirname, '..', '..', '..');
+
+const FILE_SIZE_BUDGETS = [
+  {
+    relativePath: 'server/server.ts',
+    maxLines: 4200,
+  },
+  {
+    relativePath: 'server/routes/task-board.ts',
+    maxLines: 2500,
+  },
+] as const;
+
+function lineCount(filePath: string): number {
+  return fs.readFileSync(filePath, 'utf8').split(/\r?\n/).length;
+}
+
+describe('Maintainability Guardrails', () => {
+  for (const budget of FILE_SIZE_BUDGETS) {
+    it(`${budget.relativePath} stays within ${budget.maxLines} lines`, () => {
+      const fullPath = path.join(DASHBOARD_ROOT, budget.relativePath);
+      const count = lineCount(fullPath);
+      expect(count).toBeLessThanOrEqual(budget.maxLines);
+    });
+  }
+});
